+++ mitra_validators/src/lib.rs
@@ -2,6 +2,7 @@ pub mod emojis;
 pub mod errors;
 pub mod invoices;
 pub mod media;
+pub mod oauth;
 pub mod posts;
 pub mod profiles;
 pub mod tags;