+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Validate OAuth redirect URIs.
+
 ## [2.7.0] - 2024-01-12
 
 ### Added